requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
pillow
scikit-learn
seaborn
opencv-python
streamlit
huggingface_hub